[refactor] move prompting logic to separate package
